Cancel Invitation Link Admin Only

+++ ujjwala/models.py
@@ -1743,7 +1743,8 @@ class ConnectionDisbursement(models.Model):
 	def transition_cancel_walk_in(self, *args, **kwargs):
 		self.disbursement_drive = None
 		self.walk_in_date = None
-		self.invitation.all().delete()
+		if kwargs['delete_invitation']:
+			self.invitation.all().delete()
 		self.save()
